deleted Double fields of entities. I had some problems with them.

+++ app-dao/src/main/java/com/epam/brest/task/dao/GalaxyDaoImpl.java
@@ -224,7 +224,7 @@ public class GalaxyDaoImpl implements GalaxyDao {
             galaxy.setDistance(rs.getLong("distance"));
             galaxy.setDate(rs.getDate("discoverDate"));
             galaxy.setAverageAge(rs.getLong("averageAge"));
-            galaxy.setAverageMass(rs.getDouble("averageMass"));
+            galaxy.setAverageMass(rs.getLong("averageMass"));
             return galaxy;
         }
     }
